# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.11.0 🍺
 * Do not edit manually.
 * Api
 * Sankara Eye Donation Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface Unit {
  id: number;
  name: string;
  state: string;
  district: string;
  /** @nullable */
  address?: string | null;
  coordinatorName: string;
  coordinatorWhatsapp: string;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface UnitInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  district: string;
  address?: string;
  /** @minLength 1 */
  coordinatorName: string;
  /** @minLength 10 */
  coordinatorWhatsapp: string;
  isActive?: boolean;
}

export interface UnitUpdate {
  /** @minLength 1 */
  name?: string;
  state?: string;
  district?: string;
  address?: string;
  coordinatorName?: string;
  coordinatorWhatsapp?: string;
  isActive?: boolean;
}

export interface PublicUnit {
  id: number;
  name: string;
  state: string;
  district: string;
  /** @nullable */
  address?: string | null;
}

export type AppUserRole = typeof AppUserRole[keyof typeof AppUserRole];


export const AppUserRole = {
  super_admin: 'super_admin',
  eye_bank_head: 'eye_bank_head',
  unit_coordinator: 'unit_coordinator',
} as const;

export interface AppUser {
  id: number;
  clerkId: string;
  email: string;
  name: string;
  role: AppUserRole;
  /** @nullable */
  unitId?: number | null;
  /** @nullable */
  unitName?: string | null;
  isActive: boolean;
  mustChangePassword: boolean;
  createdAt: string;
  updatedAt: string;
}

export type UserProfileRole = typeof UserProfileRole[keyof typeof UserProfileRole];


export const UserProfileRole = {
  super_admin: 'super_admin',
  eye_bank_head: 'eye_bank_head',
  unit_coordinator: 'unit_coordinator',
} as const;

export interface UserProfile {
  id: number;
  clerkId: string;
  email: string;
  name: string;
  role: UserProfileRole;
  /** @nullable */
  unitId?: number | null;
  /** @nullable */
  unitName?: string | null;
  isActive: boolean;
  mustChangePassword: boolean;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  super_admin: 'super_admin',
  eye_bank_head: 'eye_bank_head',
  unit_coordinator: 'unit_coordinator',
} as const;

export interface UserInput {
  email: string;
  /** @minLength 1 */
  name: string;
  role: UserInputRole;
  /** @nullable */
  unitId?: number | null;
  /** @minLength 8 */
  password?: string;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  super_admin: 'super_admin',
  eye_bank_head: 'eye_bank_head',
  unit_coordinator: 'unit_coordinator',
} as const;

export interface UserUpdate {
  name?: string;
  role?: UserUpdateRole;
  /** @nullable */
  unitId?: number | null;
  isActive?: boolean;
}

export interface ProfileUpdate {
  name?: string;
}

export interface PasswordReset {
  /** @minLength 8 */
  newPassword: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface ChangePasswordInput {
  oldPassword: string;
  /** @minLength 8 */
  newPassword: string;
}

export type EyeCallDonorGender = typeof EyeCallDonorGender[keyof typeof EyeCallDonorGender];


export const EyeCallDonorGender = {
  male: 'male',
  female: 'female',
  other: 'other',
} as const;

export type EyeCallStatus = typeof EyeCallStatus[keyof typeof EyeCallStatus];


export const EyeCallStatus = {
  new: 'new',
  contacted: 'contacted',
  team_sent: 'team_sent',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface EyeCall {
  id: number;
  callId: string;
  donorName: string;
  donorAge: number;
  donorGender: EyeCallDonorGender;
  timeOfDeath: string;
  causeOfDeath: string;
  referrerName: string;
  referrerMobile: string;
  referrerRelationship: string;
  state: string;
  district: string;
  pincode: string;
  address: string;
  unitId: number;
  unitName: string;
  status: EyeCallStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface EyeCallList {
  data: EyeCall[];
  total: number;
  page: number;
  limit: number;
}

export type EyeCallInputDonorGender = typeof EyeCallInputDonorGender[keyof typeof EyeCallInputDonorGender];


export const EyeCallInputDonorGender = {
  male: 'male',
  female: 'female',
  other: 'other',
} as const;

export interface EyeCallInput {
  /** @minLength 1 */
  donorName: string;
  /**
     * @minimum 0
     * @maximum 120
     */
  donorAge: number;
  donorGender: EyeCallInputDonorGender;
  timeOfDeath: string;
  causeOfDeath: string;
  /** @minLength 1 */
  referrerName: string;
  /** @minLength 10 */
  referrerMobile: string;
  referrerRelationship: string;
  state: string;
  district: string;
  pincode: string;
  address: string;
  unitId: number;
  notes?: string;
}

export type EyeCallUpdateStatus = typeof EyeCallUpdateStatus[keyof typeof EyeCallUpdateStatus];


export const EyeCallUpdateStatus = {
  new: 'new',
  contacted: 'contacted',
  team_sent: 'team_sent',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface EyeCallUpdate {
  status?: EyeCallUpdateStatus;
  notes?: string;
}

export type StatusUpdateStatus = typeof StatusUpdateStatus[keyof typeof StatusUpdateStatus];


export const StatusUpdateStatus = {
  new: 'new',
  contacted: 'contacted',
  team_sent: 'team_sent',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface StatusUpdate {
  status: StatusUpdateStatus;
  notes?: string;
}

export type PublicEyeCallInputDonorGender = typeof PublicEyeCallInputDonorGender[keyof typeof PublicEyeCallInputDonorGender];


export const PublicEyeCallInputDonorGender = {
  male: 'male',
  female: 'female',
  other: 'other',
} as const;

export interface PublicEyeCallInput {
  /** @minLength 1 */
  donorName: string;
  /**
     * @minimum 0
     * @maximum 120
     */
  donorAge: number;
  donorGender: PublicEyeCallInputDonorGender;
  timeOfDeath: string;
  causeOfDeath: string;
  /** @minLength 1 */
  referrerName: string;
  /** @minLength 10 */
  referrerMobile: string;
  referrerRelationship: string;
  state: string;
  district: string;
  pincode: string;
  address: string;
  unitId: number;
}

export interface PublicEyeCallResponse {
  eyeCall: EyeCall;
  whatsappMessage: string;
  whatsappUrl: string;
}

export interface DashboardStats {
  totalCalls: number;
  newCalls: number;
  contactedCalls: number;
  teamSentCalls: number;
  completedCalls: number;
  cancelledCalls: number;
  totalUnits: number;
  activeUnits: number;
}

export interface UnitCallCount {
  unitId: number;
  unitName: string;
  count: number;
}

export interface StatusCallCount {
  status: string;
  count: number;
}

export type ListEyeCallsParams = {
status?: ListEyeCallsStatus;
unitId?: number;
search?: string;
page?: number;
limit?: number;
};

export type ListEyeCallsStatus = typeof ListEyeCallsStatus[keyof typeof ListEyeCallsStatus];


export const ListEyeCallsStatus = {
  new: 'new',
  contacted: 'contacted',
  team_sent: 'team_sent',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export type GetRecentCallsParams = {
limit?: number;
};