# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sankara Eye Donation Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user with role and unit
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
  /auth/profile:
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Invalid email or password
        "401":
          description: Unauthorized
...
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password (used on profile or forced change)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid password or constraint error
        "401":
          description: Unauthorized

  # ─── UNITS ──────────────────────────────────────────────────────────────────
...
  /eye-calls:
    get:
      operationId: listEyeCalls
      tags: [eye-calls]
      summary: List eye donation calls (filtered by role/unit)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [new, contacted, team_sent, completed, cancelled]
        - name: unitId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of eye calls
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EyeCallList"
    post:
      operationId: createEyeCall
      tags: [eye-calls]
      summary: Create a new eye donation call (requires auth for internal, public uses separate endpoint)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EyeCallInput"
      responses:
        "201":
          description: Created eye call
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EyeCall"
        "400":
          description: Invalid input
  /eye-calls/{id}:
    get:
      operationId: getEyeCall
      tags: [eye-calls]
      summary: Get a specific eye call by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Eye call details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EyeCall"
        "404":
          description: Eye call not found
    patch:
      operationId: updateEyeCall
      tags: [eye-calls]
      summary: Update an eye call (status, notes, etc.)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EyeCallUpdate"
      responses:
        "200":
          description: Updated eye call
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EyeCall"
        "404":
          description: Eye call not found
    delete:
      operationId: deleteEyeCall
      tags: [eye-calls]
      summary: Delete an eye call (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Eye call not found
  /eye-calls/{id}/status:
    patch:
      operationId: updateEyeCallStatus
      tags: [eye-calls]
      summary: Update eye call status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Updated eye call
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EyeCall"
        "404":
          description: Eye call not found

  # ─── PUBLIC FORM ─────────────────────────────────────────────────────────────
  /public/eye-calls:
    post:
      operationId: submitPublicEyeCall
      tags: [public]
      summary: Submit public eye donation form (no auth required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicEyeCallInput"
      responses:
        "201":
          description: Submitted successfully, includes WhatsApp message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicEyeCallResponse"
        "400":
          description: Invalid input
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ─── UNIT ─────────────────────────────────────────────────────────────────
...
    UserProfile:
      type: object
      required: [id, clerkId, email, name, role, isActive, mustChangePassword]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [super_admin, eye_bank_head, unit_coordinator]
        unitId:
          type: ["integer", "null"]
        unitName:
          type: ["string", "null"]
        isActive:
          type: boolean
        mustChangePassword:
          type: boolean
...
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    ChangePasswordInput:
      type: object
      required: [oldPassword, newPassword]
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
          minLength: 8

    # ─── EYE CALL ─────────────────────────────────────────────────────────────
    EyeCall:
      type: object
      required: [id, callId, donorName, donorAge, donorGender, timeOfDeath, causeOfDeath, referrerName, referrerMobile, referrerRelationship, state, district, pincode, address, unitId, unitName, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        callId:
          type: string
        donorName:
          type: string
        donorAge:
          type: integer
        donorGender:
          type: string
          enum: [male, female, other]
        timeOfDeath:
          type: string
        causeOfDeath:
          type: string
        referrerName:
          type: string
        referrerMobile:
          type: string
        referrerRelationship:
          type: string
        state:
          type: string
        district:
          type: string
        pincode:
          type: string
        address:
          type: string
        unitId:
          type: integer
        unitName:
          type: string
        status:
          type: string
          enum: [new, contacted, team_sent, completed, cancelled]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    EyeCallList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/EyeCall"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    EyeCallInput:
      type: object
      required: [donorName, donorAge, donorGender, timeOfDeath, causeOfDeath, referrerName, referrerMobile, referrerRelationship, state, district, pincode, address, unitId]
      properties:
        donorName:
          type: string
          minLength: 1
        donorAge:
          type: integer
          minimum: 0
          maximum: 120
        donorGender:
          type: string
          enum: [male, female, other]
        timeOfDeath:
          type: string
        causeOfDeath:
          type: string
        referrerName:
          type: string
          minLength: 1
        referrerMobile:
          type: string
          minLength: 10
        referrerRelationship:
          type: string
        state:
          type: string
        district:
          type: string
        pincode:
          type: string
        address:
          type: string
        unitId:
          type: integer
        notes:
          type: string
    EyeCallUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [new, contacted, team_sent, completed, cancelled]
        notes:
          type: string
    StatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [new, contacted, team_sent, completed, cancelled]
        notes:
          type: string
    PublicEyeCallInput:
      type: object
      required: [donorName, donorAge, donorGender, timeOfDeath, causeOfDeath, referrerName, referrerMobile, referrerRelationship, state, district, pincode, address, unitId]
      properties:
        donorName:
          type: string
          minLength: 1
        donorAge:
          type: integer
          minimum: 0
          maximum: 120
        donorGender:
          type: string
          enum: [male, female, other]
        timeOfDeath:
          type: string
        causeOfDeath:
          type: string
        referrerName:
          type: string
          minLength: 1
        referrerMobile:
          type: string
          minLength: 10
        referrerRelationship:
          type: string
        state:
          type: string
        district:
          type: string
        pincode:
          type: string
        address:
          type: string
        unitId:
          type: integer
    PublicEyeCallResponse:
      type: object
      required: [eyeCall, whatsappMessage, whatsappUrl]
      properties:
        eyeCall:
          $ref: "#/components/schemas/EyeCall"
        whatsappMessage:
          type: string
        whatsappUrl:
          type: string

    # ─── DASHBOARD ────────────────────────────────────────────────────────────
...