# lib/api-zod/src/generated/types/eyeCall.ts
/**
 * Generated by orval v8.11.0 🍺
 * Do not edit manually.
 * Api
 * Sankara Eye Donation Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { EyeCallDonorGender } from './eyeCallDonorGender';
import type { EyeCallStatus } from './eyeCallStatus';

export interface EyeCall {
  id: number;
  callId: string;
  donorName: string;
  donorAge: number;
  donorGender: EyeCallDonorGender;
  timeOfDeath: string;
  causeOfDeath: string;
  referrerName: string;
  referrerMobile: string;
  referrerRelationship: string;
  state: string;
  district: string;
  pincode: string;
  address: string;
  unitId: number;
  unitName: string;
  status: EyeCallStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}