# lib/api-zod/src/generated/types/eyeCallInput.ts
/**
 * Generated by orval v8.11.0 🍺
 * Do not edit manually.
 * Api
 * Sankara Eye Donation Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { EyeCallInputDonorGender } from './eyeCallInputDonorGender';

export interface EyeCallInput {
  /** @minLength 1 */
  donorName: string;
  /**
     * @minimum 0
     * @maximum 120
     */
  donorAge: number;
  donorGender: EyeCallInputDonorGender;
  timeOfDeath: string;
  causeOfDeath: string;
  /** @minLength 1 */
  referrerName: string;
  /** @minLength 10 */
  referrerMobile: string;
  referrerRelationship: string;
  state: string;
  district: string;
  pincode: string;
  address: string;
  unitId: number;
  notes?: string;
}