# lib/api-zod/src/generated/types/eyeCallUpdate.ts
/**
 * Generated by orval v8.11.0 🍺
 * Do not edit manually.
 * Api
 * Sankara Eye Donation Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { EyeCallUpdateStatus } from './eyeCallUpdateStatus';

export interface EyeCallUpdate {
  status?: EyeCallUpdateStatus;
  notes?: string;
}